Move `takes_multiple_arguments` to `dask.utils`
